refactor: component file and folder structure
