+++ components/Header/index.tsx
@@ -1,4 +1,4 @@
-import Nav from '../Nav'
+import Nav from './Nav'
 import styles from './Header.module.css'
 
 export default function Header(): JSX.Element {
