+++ pages/index.tsx
@@ -1,7 +1,7 @@
-import Bandcamp from '../components/Bandcamp'
-import CallToAction from '../components/CallToAction'
+import Bandcamp from '../components/Player/Bandcamp'
+import CallToAction from '../components/Player/CallToAction'
 import Clouds from '../components/Clouds'
-import Cover from '../components/Cover'
+import Cover from '../components/Player/Cover'
 import Footer from '../components/Footer'
 import Header from '../components/Header'
 import Main from '../components/Main'
